Restore constants and move decimalFormat to helpers + changelog
